refactor: Extract package "golden"
